requirements.txt
Flask==2.3.3
pymongo==4.5.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
